abstracted reservation table into component
